Refactor UI and add a full rectangle struct
